feat(rust): move trezor-client build script to separate crate
update comments
