--- rust/trezor-client/protob
@@ -1,1 +1,0 @@
-../../common/protob
