+++ rust/trezor-client/src/lib.rs
@@ -12,6 +12,7 @@
 //! Please be aware that `trace` logging can contain sensitive data.
 
 #![warn(rust_2018_idioms)]
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
 mod messages;
 mod transport;
