+++ rust/trezor-client/src/messages/generated.rs
@@ -1,28 +1,3 @@
-//! This module implements the `message_type` getter for all protobuf message types.
-
-use crate::protos::{MessageType::*, *};
-
-/// Extends the protobuf Message trait to also have a static getter for the message
-/// type code.
-pub trait TrezorMessage: protobuf::Message + std::fmt::Debug {
-    const MESSAGE_TYPE: MessageType;
-
-    #[inline]
-    #[deprecated(note = "Use `MESSAGE_TYPE` instead")]
-    fn message_type() -> MessageType {
-        Self::MESSAGE_TYPE
-    }
-}
-
-/// This macro provides the TrezorMessage trait for a protobuf message.
-macro_rules! trezor_message_impl {
-	($($struct:ident => $mtype:expr),+ $(,)?) => {$(
-		impl TrezorMessage for $struct {
-			const MESSAGE_TYPE: MessageType = $mtype;
-		}
-	)+};
-}
-
 trezor_message_impl! {
     Initialize => MessageType_Initialize,
     Ping => MessageType_Ping,
@@ -66,6 +41,10 @@ trezor_message_impl! {
     FirmwareHash => MessageType_FirmwareHash,
     UnlockPath => MessageType_UnlockPath,
     UnlockedPathRequest => MessageType_UnlockedPathRequest,
+    ShowDeviceTutorial => MessageType_ShowDeviceTutorial,
+    UnlockBootloader => MessageType_UnlockBootloader,
+    AuthenticateDevice => MessageType_AuthenticateDevice,
+    AuthenticityProof => MessageType_AuthenticityProof,
     SetU2FCounter => MessageType_SetU2FCounter,
     GetNextU2FCounter => MessageType_GetNextU2FCounter,
     NextU2FCounter => MessageType_NextU2FCounter,
