+++ rust/trezor-client/src/protos/mod.rs
@@ -1,33 +1,41 @@
-include!(concat!(env!("OUT_DIR"), "/protos/mod.rs"));
+//! Bindings for Trezor protobufs.
 
+// Note: we do not use the generated `mod.rs` because we want to feature-gate some modules manually.
+// This significantly improves compile times.
+// See https://github.com/joshieDo/rust-trezor-api/pull/9 for more details.
 #[allow(ambiguous_glob_reexports)]
-pub use messages::*;
-pub use messages_bootloader::*;
-pub use messages_common::*;
-pub use messages_crypto::*;
-pub use messages_debug::*;
-pub use messages_management::*;
+mod generated {
+    macro_rules! modules {
+        ($($($feature:literal =>)? $module:ident)+) => {$(
+            $(#[cfg(feature = $feature)])?
+            mod $module;
+            $(#[cfg(feature = $feature)])?
+            pub use self::$module::*;
+        )+};
+    }
 
-macro_rules! features {
-    ($($feature:literal => $item:ident)+) => {$(
-        #[cfg(feature = $feature)]
-        #[allow(ambiguous_glob_reexports)]
-        pub use $item::*;
-    )+};
+    modules! {
+        messages
+        messages_bootloader
+        messages_common
+        messages_crypto
+        messages_debug
+        messages_management
+
+        "bitcoin" => messages_bitcoin
+        "ethereum" => messages_ethereum
+        "ethereum" => messages_ethereum_eip712
+        "ethereum" => messages_ethereum_definitions
+        "binance" => messages_binance
+        "cardano" => messages_cardano
+        "eos" => messages_eos
+        "monero" => messages_monero
+        "nem" => messages_nem
+        "ripple" => messages_ripple
+        "stellar" => messages_stellar
+        "tezos" => messages_tezos
+        "webauthn" => messages_webauthn
+    }
 }
 
-features! {
-    "bitcoin" => messages_bitcoin
-    "ethereum" => messages_ethereum
-    "ethereum" => messages_ethereum_eip712
-    "ethereum" => messages_ethereum_definitions
-    "binance" => messages_binance
-    "cardano" => messages_cardano
-    "eos" => messages_eos
-    "monero" => messages_monero
-    "nem" => messages_nem
-    "ripple" => messages_ripple
-    "stellar" => messages_stellar
-    "tezos" => messages_tezos
-    "webauthn" => messages_webauthn
-}
+pub use generated::*;
